# Finished sample/Generated code/aspnetcoreapi/src/IO.Swagger/Controllers/BooksApi.cs
/*
 * Library API
 *
 * Through this API you can access authors and their books.
 *
 * OpenAPI spec version: 1.0
 * Contact: kevin.dockx@gmail.com
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using System.Linq;
using System.Net;
using System.Threading.Tasks;
using Microsoft.AspNetCore.Http;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.WebUtilities;
using Microsoft.Extensions.Logging;
using Microsoft.Extensions.Primitives;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    public class BooksApiController : Controller
    { 
        /// <summary>
        /// Create a book for a specific author
        /// </summary>
        
        /// <param name="authorId">The id of the book author</param>
        /// <param name="body"></param>
        /// <response code="201">Success</response>
        /// <response code="400">Bad Request</response>
        /// <response code="404">Not Found</response>
        /// <response code="406">Not Acceptable</response>
        /// <response code="422">Validation error</response>
        /// <response code="500">Server Error</response>
        [HttpPost]
        [Route("/api/v1/authors/{authorId}/books")]
        [ValidateModelState]
        [SwaggerOperation("CreateBook")]
        [SwaggerResponse(statusCode: 201, type: typeof(Book), description: "Success")]
        [SwaggerResponse(statusCode: 400, type: typeof(ProblemDetails), description: "Bad Request")]
        [SwaggerResponse(statusCode: 404, type: typeof(ProblemDetails), description: "Not Found")]
        [SwaggerResponse(statusCode: 406, type: typeof(ProblemDetails), description: "Not Acceptable")]
        [SwaggerResponse(statusCode: 422, type: typeof(ValidationProblemDetails), description: "Validation error")]
        [SwaggerResponse(statusCode: 500, type: typeof(Void), description: "Server Error")]
        public virtual IActionResult CreateBook([FromRoute][Required]Guid? authorId, [FromBody]BookForCreation body)
        { 
            //TODO: Uncomment the next line to return response 201 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(201, default(Book));

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400, default(ProblemDetails));

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404, default(ProblemDetails));

            //TODO: Uncomment the next line to return response 406 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(406, default(ProblemDetails));

            //TODO: Uncomment the next line to return response 422 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(422, default(ValidationProblemDetails));

            //TODO: Uncomment the next line to return response 500 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(500, default(Void));
            string exampleJson = null;
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<Book>(exampleJson)
                        : default(Book);            //TODO: Change the data returned
            return new ObjectResult(example);
        }
        /// <summary>
        /// Create a book for a specific author
        /// </summary>
        
        /// <param name="authorId">The id of the book author</param>
        /// <param name="body"></param>
        /// <response code="201">Success</response>
        /// <response code="400">Bad Request</response>
        /// <response code="404">Not Found</response>
        /// <response code="406">Not Acceptable</response>
        /// <response code="422">Validation error</response>
        /// <response code="500">Server Error</response>
        [HttpPost]
        [Route("/api/v1/authors/{authorId}/books")]
        [ValidateModelState]
        [SwaggerOperation("CreateBook")]
        [SwaggerResponse(statusCode: 201, type: typeof(Book), description: "Success")]
        [SwaggerResponse(statusCode: 400, type: typeof(ProblemDetails), description: "Bad Request")]
        [SwaggerResponse(statusCode: 404, type: typeof(ProblemDetails), description: "Not Found")]
        [SwaggerResponse(statusCode: 406, type: typeof(ProblemDetails), description: "Not Acceptable")]
        [SwaggerResponse(statusCode: 422, type: typeof(ValidationProblemDetails), description: "Validation error")]
        [SwaggerResponse(statusCode: 500, type: typeof(Void), description: "Server Error")]
        public virtual IActionResult CreateBook([FromRoute][Required]Guid? authorId, [FromBody]BookForCreationWithAmountOfPages body)
        { 
            //TODO: Uncomment the next line to return response 201 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(201, default(Book));

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400, default(ProblemDetails));

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404, default(ProblemDetails));

            //TODO: Uncomment the next line to return response 406 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(406, default(ProblemDetails));

            //TODO: Uncomment the next line to return response 422 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(422, default(ValidationProblemDetails));

            //TODO: Uncomment the next line to return response 500 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(500, default(Void));
            string exampleJson = null;
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<Book>(exampleJson)
                        : default(Book);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Get a book by id for a specific author
        /// </summary>
        
        /// <param name="authorId">The id of the book author</param>
        /// <param name="bookId">The id of the book</param>
        /// <response code="200">Returns the requested book</response>
        /// <response code="400">Bad Request</response>
        /// <response code="404">Not Found</response>
        /// <response code="406">Not Acceptable</response>
        /// <response code="500">Server Error</response>
        [HttpGet]
        [Route("/api/v1/authors/{authorId}/books/{bookId}")]
        [ValidateModelState]
        [SwaggerOperation("GetBook")]
        [SwaggerResponse(statusCode: 200, type: typeof(Book), description: "Returns the requested book")]
        [SwaggerResponse(statusCode: 400, type: typeof(ProblemDetails), description: "Bad Request")]
        [SwaggerResponse(statusCode: 404, type: typeof(ProblemDetails), description: "Not Found")]
        [SwaggerResponse(statusCode: 406, type: typeof(ProblemDetails), description: "Not Acceptable")]
        [SwaggerResponse(statusCode: 500, type: typeof(Void), description: "Server Error")]
        public virtual IActionResult GetBook([FromRoute][Required]Guid? authorId, [FromRoute][Required]Guid? bookId)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(Book));

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400, default(ProblemDetails));

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404, default(ProblemDetails));

            //TODO: Uncomment the next line to return response 406 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(406, default(ProblemDetails));

            //TODO: Uncomment the next line to return response 500 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(500, default(Void));
            string exampleJson = null;
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<Book>(exampleJson)
                        : default(Book);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Get the books for a specific author
        /// </summary>
        
        /// <param name="authorId">The id of the book author</param>
        /// <response code="200">Success</response>
        /// <response code="400">Bad Request</response>
        /// <response code="404">Not Found</response>
        /// <response code="406">Not Acceptable</response>
        /// <response code="500">Server Error</response>
        [HttpGet]
        [Route("/api/v1/authors/{authorId}/books")]
        [ValidateModelState]
        [SwaggerOperation("GetBooks")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<Book>), description: "Success")]
        [SwaggerResponse(statusCode: 400, type: typeof(ProblemDetails), description: "Bad Request")]
        [SwaggerResponse(statusCode: 404, type: typeof(ProblemDetails), description: "Not Found")]
        [SwaggerResponse(statusCode: 406, type: typeof(ProblemDetails), description: "Not Acceptable")]
        [SwaggerResponse(statusCode: 500, type: typeof(Void), description: "Server Error")]
        public virtual IActionResult GetBooks([FromRoute][Required]Guid? authorId)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<Book>));

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400, default(ProblemDetails));

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404, default(ProblemDetails));

            //TODO: Uncomment the next line to return response 406 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(406, default(ProblemDetails));

            //TODO: Uncomment the next line to return response 500 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(500, default(Void));
            string exampleJson = null;
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<List<Book>>(exampleJson)
                        : default(List<Book>);            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}