# Finished sample/Generated code/aspnetcoreapi/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Library API",
    "description" : "Through this API you can access authors and their books.",
    "contact" : {
      "name" : "Kevin Dockx",
      "url" : "https://www.twitter.com/KevinDockx",
      "email" : "kevin.dockx@gmail.com"
    },
    "license" : {
      "name" : "MIT License",
      "url" : "https://opensource.org/licenses/MIT"
    },
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/api/v1/authors/{authorId}/books" : {
      "get" : {
        "tags" : [ "Books" ],
        "summary" : "Get the books for a specific author",
        "operationId" : "GetBooks",
        "parameters" : [ {
          "name" : "authorId",
          "in" : "path",
          "description" : "The id of the book author",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "uniqueItems" : false,
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Book"
                  }
                }
              },
              "application/xml" : {
                "schema" : {
                  "uniqueItems" : false,
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Book"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "406" : {
            "description" : "Not Acceptable",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500" : {
            "description" : "Server Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Void"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Void"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Books" ],
        "summary" : "Create a book for a specific author",
        "operationId" : "CreateBook",
        "parameters" : [ {
          "name" : "authorId",
          "in" : "path",
          "description" : "The id of the book author",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BookForCreation"
              }
            },
            "application/vnd.marvin.bookforcreation+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BookForCreation"
              }
            },
            "application/vnd.marvin.bookforcreationwithamountofpages+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BookForCreationWithAmountOfPages"
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Book"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Book"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "406" : {
            "description" : "Not Acceptable",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422" : {
            "description" : "Validation error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "500" : {
            "description" : "Server Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Void"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Void"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/authors/{authorId}/books/{bookId}" : {
      "get" : {
        "tags" : [ "Books" ],
        "summary" : "Get a book by id for a specific author",
        "operationId" : "GetBook",
        "parameters" : [ {
          "name" : "authorId",
          "in" : "path",
          "description" : "The id of the book author",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "bookId",
          "in" : "path",
          "description" : "The id of the book",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns the requested book",
            "content" : {
              "application/vnd.marvin.book+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Book"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Book"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Book"
                }
              },
              "application/vnd.marvin.bookwithconcatenatedauthorname+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BookWithConcatenatedAuthorName"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "content" : {
              "application/vnd.marvin.book+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found",
            "content" : {
              "application/vnd.marvin.book+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "406" : {
            "description" : "Not Acceptable",
            "content" : {
              "application/vnd.marvin.book+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500" : {
            "description" : "Server Error",
            "content" : {
              "application/vnd.marvin.book+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Void"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Void"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Void"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "ProblemDetails" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "detail" : {
            "type" : "string"
          },
          "instance" : {
            "type" : "string"
          }
        },
        "additionalProperties" : {
          "type" : "object"
        }
      },
      "Void" : {
        "type" : "object"
      },
...
      "ValidationProblemDetails" : {
        "type" : "object",
        "properties" : {
          "errors" : {
            "type" : "object",
            "additionalProperties" : {
              "uniqueItems" : false,
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            },
            "readOnly" : true
          },
          "type" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "detail" : {
            "type" : "string"
          },
          "instance" : {
            "type" : "string"
          }
        },
        "additionalProperties" : {
          "type" : "object"
        }
      },
...
      "Book" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "authorFirstName" : {
            "type" : "string"
          },
          "authorLastName" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          }
        }
      },
      "BookForCreation" : {
        "required" : [ "description", "title" ],
        "type" : "object",
        "properties" : {
          "title" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          }
        }
      },
      "BookForCreationWithAmountOfPages" : {
        "required" : [ "description", "title" ],
        "type" : "object",
        "properties" : {
          "amountOfPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "title" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          }
        }
      },
      "BookWithConcatenatedAuthorName" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "author" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          }
        }
      }
    }
  }
}